requirements.txt
dj-database-url==0.5.0
Django==3.2
django-phonenumber-field==5.2.0
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.19.2
phonenumbers==8.12.27
Pillow==6.2.1